// tmp_app/Api.ts
export interface input$api$videos$upload {
    title: string, description: string, keywords: string, category: string, extension: string, thumbnail_png: string
}